+++ CHANGELOG.md
@@ -1,10 +1,39 @@
-# Tip
+# Unreleased
+
+## Fixes
+
+* Fix *ConnPool.Deallocate() not deleting prepared statement from map
+
+## Features
+
+* Encode and decode between all Go and PostgreSQL integer types with bounds checking
+* Decode inet/cidr to net.IP
+* Encode/decode [][]byte to/from bytea[]
+
+## Performance
+
+* Substantial reduction in memory allocations
+
+# 2.8.1 (March 24, 2016)
+
+## Features
+
+* Scan accepts nil argument to ignore a column
+
+## Fixes
+
+* Fix compilation on 32-bit architecture
+* Fix Tx.status not being set on error on Commit
+* Fix Listen/Unlisten with special characters
+
+# 2.8.0 (March 18, 2016)
 
 ## Fixes
 
 * Fix unrecognized commit failure
 * Fix msgReader.rxMsg bug when msgReader already has error
 * Go float64 can no longer be encoded to a PostgreSQL float4
+* Fix connection corruption when query with error is closed early
 
 ## Features
 
@@ -28,6 +57,8 @@ standard database/sql package such as
 * Add Decode() to allow customer Decoders to reuse internal decoding functionality
 * Add ConnPool.Prepare method
 * Add ConnPool.Deallocate method
+* Add Scan to uint32 and uint64 (utrack)
+* Add encode and decode to []uint16, []uint32, and []uint64 (Max Musatov)
 
 ## Performance